add 数据字典缓存转换

+++ learn-api/src/main/java/io/github/suxil/service/domain/cdm/CdmDict.java
@@ -38,7 +38,7 @@ public class CdmDict extends BaseDomain {
 
     @ApiModelProperty(name = "seq", value = "序号")
     @Column(name = "seq")
-    private BigDecimal seq;
+    private Integer seq;
 
     @ApiModelProperty(name = "categoryId", value = "类型id")
     @Column(name = "category_id")
